Do that brew

+++ .config/oh-my-zsh/plugins/bubba-brew/bubba-brew.plugin.zsh
@@ -16,19 +16,31 @@ if [ ! -f  $brew_bin ]; then
     curl -L https://github.com/Homebrew/brew/tarball/master | tar xz --strip 1 -C $brew_prefix
     mkdir -p \
         $brew_prefix/etc \
-        $brew_prefix//include \
-        $brew_prefix//lib \
-        $brew_prefix//opt \
-        $brew_prefix//sbin \
-        $brew_prefix//share \
-        $brew_prefix//var/homebrew/linked \
-        $brew_prefix//Cellar
+        $brew_prefix/include \
+        $brew_prefix/lib \
+        $brew_prefix/opt \
+        $brew_prefix/sbin \
+        $brew_prefix/share \
+        $brew_prefix/var/homebrew/linked \
+        $brew_prefix/Cellar
 fi
 
 export PATH=/Users/bubba/homebrew/bin:$PATH
-brew install coreutils
+
+if [ ! -f  $brew_prefix/bin/gdircolors ]; then
+    brew install coreutils
+fi
+
+if [ ! -f  $brew_prefix/bin/pyenv ]; then
+    brew install pyenv
+    pyenv init
+fi
 
 
+if [ ! -f  $HOME/bin/phpbrew ]; then
+    curl -L -o $HOME/bin/phpbrew https://github.com/phpbrew/phpbrew/raw/master/phpbrew
+    chmod +x $HOME/bin/phpbrew
+    phpbrew init
+fi
 
-curl -L -O=$HOME/bin/phpbrew https://github.com/phpbrew/phpbrew/raw/master/phpbrew
-chmod +x $HOME/bin/phpbrew
+
